Create README.md
CUB3D
This project is inspired by the world-famous eponymous 90’s game, which
was the first FPS ever. It was to explore ray-casting. My goal was to
make a dynamic view inside a maze, in which you’ll have to find the way.


To start the game:
make
./cub3D map.cub

To make the screenshot in bmp format of the first image:

make
./cub3D map.cub --save